+++ docs/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 Things under "main" are not released yet.
 
+## main
+
+- Allow prefixing a var with `_` to silence the unused variable warning.
+
 ## v0.9.7
 
 Add types for the following functions: